Refactor materialized views package
